services/market-data-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
finnhub-python>=2.4.19
websocket-client>=1.7.0
yfinance>=0.2.65
python-dotenv>=1.0.1
asyncpg>=0.29.0
pydantic-settings>=2.0.0
scipy>=1.11.0
numpy>=1.24.0
